% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaned_realtordata.R
\docType{data}
\name{cleaned_realtordata}
\alias{cleaned_realtordata}
\title{Cleaned realtor data}
\format{
4 columns with price - in dollars, bed, bath, house_size - in sq ft
}
\usage{
data(cleaned_realtordata)
}
\description{
realtor data set loaded to create train/test split for model fits
}
\details{
realtor dataset without NAs
}
\keyword{datasets}
