% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lm.R
\name{predict_lm}
\alias{predict_lm}
\title{Function that predicts home price using linear model}
\usage{
predict_lm(nor_User_data)
}
\arguments{
\item{x}{Normalized user data in a dataframe}
}
\value{
unscaled predicted price
}
\description{
Function that predicts home price using linear model
}
\examples{
predictedPrice <- perdict_lm(normalized_User_data)
}
