% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nn.R
\name{predict_nn}
\alias{predict_nn}
\title{Function that predicts home price using neural net}
\usage{
predict_nn(nor_User_data)
}
\arguments{
\item{x}{Normalized user data in a dataframe}
}
\value{
unscaled predicted price
}
\description{
Function that predicts home price using neural net
}
\examples{
predictedPrice <- predict_nn(normalized_User_data)
}
