% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealHomePrice .R
\docType{package}
\name{RealHomePrice}
\alias{RealHomePrice}
\title{RealHomePrice Package}
\description{
Contains a function for predicting home price with neural net and MLR.
}
\author{
Moreen Owino, Bipin Parajuli, Russell Marvin, Nicholas Scholl
}
