% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data.R
\docType{data}
\name{train_data}
\alias{train_data}
\title{train_data}
\format{
4 columns with price - in dollars, bed, bath, house_size - in sq ft
}
\usage{
data(train_data)
}
\description{
80\% split of data - used to train models
}
\details{
subset of cleaned_realtordata
}
\keyword{datasets}
