% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale_price.R
\name{unscale_price}
\alias{unscale_price}
\title{Function that unscales the predicted price}
\usage{
unscale_price(price)
}
\arguments{
\item{x}{predicted price by the model}
}
\value{
unscaled prediction
}
\description{
Function that unscales the predicted price
}
\examples{
unscaled_price <- unscale_price(pricelm)
}
