% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_input.R
\name{normalize_input}
\alias{normalize_input}
\title{Function that normalizes the userinput using -min/max}
\usage{
normalize_input(x)
}
\arguments{
\item{x}{user data in a dataframe, numeric values}
}
\value{
dataframe of normalized user input
}
\description{
Function that normalizes the userinput using -min/max
}
