% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lm.R
\name{predict_lm}
\alias{predict_lm}
\title{Function that predicts home price using linear model}
\usage{
predict_lm(x)
}
\arguments{
\item{x}{Normalized user data in a dataframe, numeric values}
}
\value{
unscaled predicted price
}
\description{
Function that predicts home price using linear model
}
