% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Price.R
\name{predict_Price}
\alias{predict_Price}
\title{Function that combines all the other functions for a human readable result}
\usage{
predict_Price(user_data)
}
\arguments{
\item{x}{user data in a dataframe}
}
\value{
a dataframe with predicted prices
}
\description{
Function that combines all the other functions for a human readable result
}
\examples{
 predict_Price(User_data)
}
