% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealHomePrice.R
\docType{package}
\name{RealHomePrice}
\alias{RealHomePrice}
\title{RealHomePrice Package}
\description{
Contains a function for predicting home price with neural net and MLR.
}
\details{
References: https://datascienceplus.com/fitting-neural-network-in-r/,
https://www.nature.com/articles/nature14539, https://towardsdatascience.com/introduction-to-math-behind-neural-networks-e8b60dbbdeba,
https://en.wikipedia.org/wiki/Activation_function
}
\author{
Moreen Owino, Bipin Parajuli, Russell Marvin, Nicholas Scholl
}
